requirements.txt
mcp[cli]
google-generativeai
python-dotenv
httpx